# infra/contracts/product-service.openapi.yaml
openapi: 3.0.3
info:
  title: product-service
  description: >
    Catalog service — brands, categories, products, variants, images.
    Code-first contract (see infra/docs/adr/0001-contract-sharing-mechanism.md and the
    Sprint 1 plan's Q7): written from the real NestJS implementation in
    `product-service/src`, not the other way around. Documents GET/POST/PATCH/DELETE
    /products, POST /products/:id/images, full CRUD for /categories and /brands, and GET /health
    (admin product management pass, see .planning/2026-07-11-admin-product-management.md).
    Face shapes are stored as a product enum array; variant stock is stored on the variant.
  version: "0.2.0"
servers:
  - url: http://localhost:3002
    description: product-service, direct (local dev only)
  - url: http://localhost:8080/api
    description: via api-gateway (same shapes, `/api` prefix — see api-gateway proxy)

# Read endpoints are public. Brand/category write endpoints require `catalog:manage`
# when requested through api-gateway; product-service trusts that edge enforcement.
security: []
...
paths:
...
  /products:
    get:
      tags: [products]
      summary: List products (filter + paginate)
      operationId: listProducts
      parameters:
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
        - name: brandId
          in: query
          schema:
            type: string
            format: uuid
        - name: frameShape
          in: query
          schema:
            $ref: "#/components/schemas/FrameShape"
        - name: genderTarget
          in: query
          schema:
            $ref: "#/components/schemas/GenderTarget"
        - name: faceShape
          in: query
          description: >
            Filters to products whose `faceShapes` array includes this face shape.
          schema:
            $ref: "#/components/schemas/FaceShape"
        - name: status
          in: query
          description: >
            Defaults to `PUBLISHED` server-side when omitted (unless `includeAllStatuses`
            is set).
          schema:
            $ref: "#/components/schemas/ProductStatus"
        - name: includeAllStatuses
          in: query
          description: >
            Admin-only escape hatch (AC1): when `true` and `status` is not given, skips
            the default `status = PUBLISHED` filter and returns products in any status.
            Ignored if `status` is set. Public storefront behavior (default-to-PUBLISHED)
            is unchanged when this is omitted.
          schema:
            type: boolean
        - name: minPrice
          in: query
          schema:
            type: number
            minimum: 0
            exclusiveMinimum: true
        - name: maxPrice
          in: query
          schema:
            type: number
            minimum: 0
            exclusiveMinimum: true
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Paginated list of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedProductResponse"
        "400":
          description: Invalid query parameters (validation failure)

    post:
      tags: [products]
      summary: Create a product (admin)
      description: >
        `slug` and `sku` are auto-generated server-side from `name` (never supplied by
        the caller). `status` defaults to `PUBLISHED` when omitted (not the DB's own
        `DRAFT` default) since the admin UI's toggle only ever represents
        published/archived.
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "400":
          description: >
            Validation failure, or `categoryId`/`brandId` does not reference an
            existing row.
  /products/{id}:
    get:
      tags: [products]
      summary: Get a single product by id
      operationId: getProductById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "404":
          description: Product not found

    patch:
      tags: [products]
      summary: Update a product (admin)
      description: >
        Partial update — only fields present in the body are changed. If `faceShapes`
        is included, it replaces the product's full face-shape tag set.
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "400":
          description: >
            Validation failure, or `categoryId`/`brandId` does not reference an
            existing row.
        "404":
          description: Product not found

    delete:
      tags: [products]
      summary: Soft-delete a product (admin)
      description: >
        Sets `deleted_at` (TypeORM soft delete) — never a hard delete. The row is
        excluded from all subsequent list/detail queries once deleted.
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Product soft-deleted
        "404":
          description: Product not found
  /products/{id}/images:
    post:
      tags: [products]
      summary: Upload a product or variant image (admin)
      description: >
        Multipart upload. Appends to the image list rather than replacing anything —
        `sortOrder` is the current max within the target group + 1. Omitting `variantId`
        attaches the image to the base product (max 8 images); passing a `variantId`
        attaches it to that variant instead (max 5 images), and the variant must belong
        to this product. A newly uploaded image never starts as the thumbnail — the
        admin sets it explicitly via `PATCH .../images/{imageId}`.
        Rejects anything that isn't `image/jpeg`, `image/png`, or `image/webp`, or is
        over 5 MB, with a 400 (server-side check, not just a client-side hint).
      operationId: uploadProductImage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                variantId:
                  type: string
                  format: uuid
                  description: Omit to attach the image to the base product.
              required: [file]
      responses:
        "201":
          description: Image uploaded and attached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductImageResponse"
        "400":
          description: >
            Missing file, disallowed content type, file over 5 MB, `variantId` not
            belonging to this product, or the group's image count limit (8 for the
            product, 5 for a variant) already reached.
        "404":
          description: Product not found
...
  /categories:
    get:
      tags: [categories]
      summary: List all categories
      operationId: listCategories
      responses:
        "200":
          description: Public flat list of one-level categories. This remains an array, not a pagination envelope.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryResponse"

    post:
      tags: [categories]
      summary: Create a category (catalog manager)
      description: >
        Through api-gateway, requires a bearer token with `catalog:manage`.
        The server creates a unique slug from the name.
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /categories/{id}:
    get:
      tags: [categories]
      summary: Get a category by id
      operationId: getCategoryById
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Category found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "404":
          $ref: "#/components/responses/NotFoundError"

    patch:
      tags: [categories]
      summary: Update a category (catalog manager)
      description: >
        Through api-gateway, requires a bearer token with `catalog:manage`. Changing
        the name regenerates the unique slug.
      operationId: updateCategory
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryRequest"
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"

    delete:
      tags: [categories]
      summary: Delete a category (catalog manager)
      description: >
        Through api-gateway, requires a bearer token with `catalog:manage`. A category
        referenced by a product cannot be deleted until its products are moved or deleted.
      operationId: deleteCategory
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "204":
          description: Category deleted
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          $ref: "#/components/responses/CategoryInUseError"
  /brands:
    get:
      tags: [brands]
      summary: List all brands
      operationId: listBrands
      responses:
        "200":
          description: Public flat list of brands. This remains an array, not a pagination envelope.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BrandResponse"

    post:
      tags: [brands]
      summary: Create a brand (catalog manager)
      description: >
        Through api-gateway, requires a bearer token with `catalog:manage`. Brand names
        are unique after trimming whitespace and case normalization.
      operationId: createBrand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBrandRequest"
      responses:
        "201":
          description: Brand created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "409":
          $ref: "#/components/responses/BrandNameConflictError"
  /brands/{id}:
    get:
      tags: [brands]
      summary: Get a brand by id
      operationId: getBrandById
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Brand found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandResponse"
        "404":
          $ref: "#/components/responses/NotFoundError"

    patch:
      tags: [brands]
      summary: Update a brand (catalog manager)
      description: Through api-gateway, requires a bearer token with `catalog:manage`.
      operationId: updateBrand
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBrandRequest"
      responses:
        "200":
          description: Brand updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          $ref: "#/components/responses/BrandNameConflictError"

    delete:
      tags: [brands]
      summary: Delete a brand (catalog manager)
      description: >
        Through api-gateway, requires a bearer token with `catalog:manage`. A brand
        referenced by a product cannot be deleted until its products are moved or deleted.
      operationId: deleteBrand
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "204":
          description: Brand deleted
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          $ref: "#/components/responses/BrandInUseError"
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string
      required: [statusCode, message]
    FrameShape:
      type: string
      enum: [ROUND, SQUARE, OVAL, CAT_EYE, AVIATOR, RECTANGLE, WAYFARER, RIMLESS]
    GenderTarget:
      type: string
      enum: [MALE, FEMALE, UNISEX]
    ProductStatus:
      type: string
      enum: [DRAFT, PUBLISHED, ARCHIVED]
    FaceShape:
      type: string
      description: >
        Which face shapes a product suits (FR7).
      enum: [ROUND, SQUARE, OVAL, HEART, DIAMOND, OBLONG]
    BrandSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        logoUrl:
          type: string
          nullable: true
      required: [id, name, logoUrl]
    BrandResponse:
      type: object
      description: Public and management representation of a brand.
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        logoUrl:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
      required: [id, name, logoUrl, description]
    CreateBrandRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        logoUrl:
          type: string
          format: uri
          maxLength: 512
          nullable: true
        description:
          type: string
          nullable: true
      required: [name]
    UpdateBrandRequest:
      type: object
      description: Partial update — every field from CreateBrandRequest is optional.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        logoUrl:
          type: string
          format: uri
          maxLength: 512
          nullable: true
        description:
          type: string
          nullable: true
    CategorySummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
      required: [id, name, slug]
    CategoryResponse:
      type: object
      description: Public and management representation of a one-level category.
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
      required: [id, name, slug]
    CreateCategoryRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
      required: [name]
    UpdateCategoryRequest:
      type: object
      description: Partial update — the server regenerates slug when name changes.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
    ProductVariantResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        color:
          type: string
        colorHex:
          type: string
          nullable: true
          description: >
            Hex color code (e.g. `#E53935`), or `null` for a legacy variant created
            before the color picker was added.
        size:
          type: string
        extraPrice:
          type: number
        skuVariant:
          type: string
        stock:
          type: integer
          description: >
            Available stock (`quantity - reservedQuantity`). Present only in
            `GET /products/:id`; omitted from catalog listing responses.
      required: [id, color, colorHex, size, extraPrice, skuVariant]
    ProductImageResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        variantId:
          type: string
          format: uuid
          nullable: true
        imageUrl:
          type: string
        isThumbnail:
          type: boolean
        sortOrder:
          type: integer
      required: [id, variantId, imageUrl, isThumbnail, sortOrder]
    ProductResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        faceFitNote:
          type: string
          nullable: true
          description: Free-text note on which faces this product fits well (FR7).
        frameShape:
          $ref: "#/components/schemas/FrameShape"
        genderTarget:
          $ref: "#/components/schemas/GenderTarget"
        material:
          type: string
          nullable: true
        basePrice:
          type: number
        status:
          $ref: "#/components/schemas/ProductStatus"
        brand:
          $ref: "#/components/schemas/BrandSummary"
        category:
          $ref: "#/components/schemas/CategorySummary"
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariantResponse"
        images:
          type: array
          items:
            $ref: "#/components/schemas/ProductImageResponse"
        faceShapes:
          type: array
          items:
            $ref: "#/components/schemas/FaceShape"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - sku
        - name
        - slug
        - description
        - faceFitNote
        - frameShape
        - genderTarget
        - material
        - basePrice
        - status
        - brand
        - category
        - variants
        - images
        - faceShapes
        - createdAt
        - updatedAt
    CreateProductRequest:
      type: object
      description: >
        `sku`/`slug` are never accepted here — they're auto-generated server-side
        (Q1). `status` is optional; defaults to `PUBLISHED` if omitted.
      properties:
        name:
          type: string
          maxLength: 255
        categoryId:
          type: string
          format: uuid
        brandId:
          type: string
          format: uuid
        frameShape:
          $ref: "#/components/schemas/FrameShape"
        genderTarget:
          $ref: "#/components/schemas/GenderTarget"
        material:
          type: string
          nullable: true
        basePrice:
          type: number
          exclusiveMinimum: true
          minimum: 0
        description:
          type: string
          nullable: true
        faceFitNote:
          type: string
          nullable: true
        faceShapes:
          type: array
          items:
            $ref: "#/components/schemas/FaceShape"
        status:
          $ref: "#/components/schemas/ProductStatus"
      required:
        - name
        - categoryId
        - brandId
        - frameShape
        - genderTarget
        - basePrice
    UpdateProductRequest:
      type: object
      description: Partial update — every field from CreateProductRequest is optional.
      properties:
        name:
          type: string
          maxLength: 255
        categoryId:
          type: string
          format: uuid
        brandId:
          type: string
          format: uuid
        frameShape:
          $ref: "#/components/schemas/FrameShape"
        genderTarget:
          $ref: "#/components/schemas/GenderTarget"
        material:
          type: string
          nullable: true
        basePrice:
          type: number
          exclusiveMinimum: true
          minimum: 0
        description:
          type: string
          nullable: true
        faceFitNote:
          type: string
          nullable: true
        faceShapes:
          type: array
          items:
            $ref: "#/components/schemas/FaceShape"
        status:
          $ref: "#/components/schemas/ProductStatus"
    PaginatedProductResponse:
      type: object
      description: >
        Flat pagination envelope (NOT a nested data/meta shape) — see
        product-service/src/routes/dto/paginated-response.dto.ts.
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ProductResponse"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [items, total, page, limit, totalPages]
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    ValidationError:
      description: Dữ liệu gửi lên không hợp lệ.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnauthorizedError:
      description: Cần đăng nhập để thực hiện thao tác này.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ForbiddenError:
      description: Bạn không có quyền quản lý danh mục.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFoundError:
      description: Không tìm thấy thương hiệu hoặc danh mục.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BrandNameConflictError:
      description: Tên thương hiệu đã tồn tại.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BrandInUseError:
      description: Không thể xoá thương hiệu vì vẫn còn sản phẩm đang sử dụng. Vui lòng chuyển hoặc xoá sản phẩm trước.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    CategoryInUseError:
      description: Không thể xoá danh mục vì vẫn còn sản phẩm đang sử dụng. Vui lòng chuyển hoặc xoá sản phẩm trước.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"